my_requirements.txt
spacy
clean-text
rank-bm25
flask
pyserini
unidecode
gdown
gym
selenium
thefuzz
textworld
openai
python-dotenv
backoff
